+++ spec/elastic_apm_spec.rb
@@ -176,7 +176,7 @@ RSpec.describe ElasticAPM do
       should delegate :report_message,
         to: agent, args: ['NOT OK', { backtrace: Array, context: nil }]
     end
-    it { should delegate :set_tag, to: agent, args: [nil, nil] }
+    it { should delegate :set_label, to: agent, args: [nil, nil] }
     it { should delegate :set_custom_context, to: agent, args: [nil] }
     it { should delegate :set_user, to: agent, args: [nil] }