+++ spec/integration/opentracing_spec.rb
@@ -152,7 +152,7 @@ RSpec.describe 'OpenTracing bridge', :intercept do
     it 'traces nested spans' do
       OpenTracing.start_active_span(
         'operation_name',
-        tags: { test: '0' }
+        labels: { test: '0' }
       ) do |scope|
         expect(scope).to be_a(ElasticAPM::OpenTracing::Scope)
         expect(OpenTracing.active_span).to be scope.span
@@ -160,7 +160,7 @@ RSpec.describe 'OpenTracing bridge', :intercept do
 
         OpenTracing.start_active_span(
           'nested',
-          tags: { test: '1' }
+          labels: { test: '1' }
         ) do |nested_scope|
           expect(OpenTracing.active_span).to_not be_nil
           expect(nested_scope.span).to eq OpenTracing.active_span
@@ -175,10 +175,10 @@ RSpec.describe 'OpenTracing bridge', :intercept do
       expect(@intercepted.spans.length).to be 2
 
       transaction, = @intercepted.transactions
-      expect(transaction.context.tags).to match(test: '0')
+      expect(transaction.context.labels).to match(test: '0')
 
       span = @intercepted.spans.last
-      expect(span.context.tags).to match(test: '1')
+      expect(span.context.labels).to match(test: '1')
     end
   end
 
@@ -207,7 +207,7 @@ RSpec.describe 'OpenTracing bridge', :intercept do
       end
     end
 
-    describe 'set_tag' do
+    describe 'set_label' do
       subject { described_class.new(elastic_span, trace_context) }
 
       shared_examples :opengraph_span do
@@ -216,10 +216,10 @@ RSpec.describe 'OpenTracing bridge', :intercept do
           expect(elastic_span.name).to eq 'Test'
         end
 
-        describe 'set_tag' do
-          it 'sets tag' do
-            subject.set_tag :custom_key, 'custom_type'
-            expect(subject.elastic_span.context.tags[:custom_key])
+        describe 'set_label' do
+          it 'sets label' do
+            subject.set_label :custom_key, 'custom_type'
+            expect(subject.elastic_span.context.labels[:custom_key])
               .to eq 'custom_type'
           end
         end
@@ -232,10 +232,10 @@ RSpec.describe 'OpenTracing bridge', :intercept do
         it_behaves_like :opengraph_span
 
         it 'knows user fields' do
-          subject.set_tag 'user.id', 1
-          subject.set_tag 'user.username', 'someone'
-          subject.set_tag 'user.email', 'someone@example.com'
-          subject.set_tag 'user.other_field', 'someone@example.com'
+          subject.set_label 'user.id', 1
+          subject.set_label 'user.username', 'someone'
+          subject.set_label 'user.email', 'someone@example.com'
+          subject.set_label 'user.other_field', 'someone@example.com'
 
           user = subject.elastic_span.context.user
           expect(user.id).to eq 1
@@ -255,7 +255,7 @@ RSpec.describe 'OpenTracing bridge', :intercept do
         it_behaves_like :opengraph_span
 
         it "doesn't explode on user fields" do
-          expect { subject.set_tag 'user.id', 1 }
+          expect { subject.set_label 'user.id', 1 }
             .to_not raise_error
         end
       end